refined the masking technique of ViWordFormer

+++ models/viwordformer/viwordformer_seq_label.py
@@ -94,7 +94,7 @@ class PhrasalLexemeEncoderLayer(nn.Module):
         _, self_attn = self.self_attn(inputs, inputs, inputs, self_attention_mask, average_attn_weights=False)
 
         # performing phrasal lexeme attention
-        attention_mask = 1 - attention_mask
+        attention_mask = (1 - attention_mask).long()
         P, phrasal_attn = self.phrasal_lexeme_attn(inputs, attention_mask, phrasal_attn)
 
         attn_scores = P * self_attn
